factorisation dashboard.vue add fetchUserBookings() to rooms.js
